requirements.txt
ffpyplayer
moviepy
SpeechRecognition
ffmpeg-python
opencv-python